address remaining PR #5526 review feedback
- Fix stale `owners/<user>` wording in tests/support/reborn/CLAUDE.md and
  scenario_two_actors_own_threads.rs to describe the real physical layout
  (/tenants/<tenant>/users/<user>/threads).
- Move the session_thread.rs unit tests for path_segment/threads_mount_view
  out of a path-included `#[cfg(test)] mod tests` (compiled+run once per
  Reborn integration binary) into tests/support_unit_tests.rs, which compiles
  and runs support-module tests exactly once. Widen path_segment and
  threads_mount_view to pub(crate) so the moved tests can call them.
- Drop the "(review finding, PR #5526)" review-metadata aside from the
  moved path_segment test's doc comment.
