+++ tests/reborn_group_multiuser/scenario_two_actors_own_threads.rs
@@ -62,7 +62,8 @@ pub async fn run(g: &RebornIntegrationGroup) -> HarnessResult<()> {
 /// separation, not just "didn't crash" — the banana-99 pattern):
 /// `reader`'s own thread must not surface `other`'s reply, and `other`'s
 /// thread must not be readable through `reader`'s owner scope at all — each
-/// owner's records live under a separate `owners/<user>` subtree. One helper
+/// owner's records live under a separate `/tenants/<tenant>/users/<user>/threads`
+/// subtree. One helper
 /// called once per direction, so the symmetric check cannot silently de-sync.
 async fn assert_cannot_read_other_actor(
     reader: &RebornIntegrationHarness,
